% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_functions.R
\name{calculate_federal_income_tax}
\alias{calculate_federal_income_tax}
\title{Calculate Federal Income Tax Deductions and Taxable Income}
\usage{
calculate_federal_income_tax(df, federal_standard_deduction)
}
\arguments{
\item{df}{Dataframe with starting_income, household_type, health_ins_premium,
and (optionally) annual_ss_ssdi_benefit -- see calculate_ss_benefit_taxability()}

\item{federal_standard_deduction}{Dataframe with standard deductions by filing status}
}
\value{
Dataframe with deduction and taxable income columns added
}
\description{
Calculate Federal Income Tax Deductions and Taxable Income
}
